Make Cripts buildable with BoringSSL

+++ src/cripts/Crypto.cc
@@ -170,7 +170,7 @@ Crypto::detail::Cipher::_initialize()
 
   TSAssert(_ctx == nullptr);
   TSReleaseAssert(_cipher != nullptr);
-  TSReleaseAssert(_key_len == EVP_CIPHER_key_length(_cipher)); // Make sure the crypto key was correct size
+  TSReleaseAssert(_key_len == static_cast<int>(EVP_CIPHER_key_length(_cipher))); // Make sure the crypto key was correct size
 
   memset(iv, 0, sizeof(iv)); // The IV is always '0x0'
   _ctx = EVP_CIPHER_CTX_new();
